Creating a file for each separated task class
